requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
bcrypt==4.1.2
pyjwt==2.8.0
python-multipart==0.0.6
